refactor installer module
